Refactor image create command (closes #818)
* Move providers images and instances methods to dedicated files
