+++ cmd/image.go
@@ -11,6 +11,49 @@ import (
 	"github.com/spf13/cobra"
 )
 
+// ImageCommands provides image related command on GCP
+func ImageCommands() *cobra.Command {
+	var config, targetCloud, zone string
+	var cmdImage = &cobra.Command{
+		Use:       "image",
+		Short:     "manage nanos images",
+		ValidArgs: []string{"create", "list", "delete", "resize", "sync"},
+		Args:      cobra.OnlyValidArgs,
+	}
+	cmdImage.PersistentFlags().StringVarP(&config, "config", "c", "", "ops config file")
+	cmdImage.PersistentFlags().StringVarP(&targetCloud, "target-cloud", "t", "onprem", "cloud platform [gcp, aws, do, vultr, onprem]")
+	cmdImage.PersistentFlags().StringVarP(&zone, "zone", "z", os.Getenv("GOOGLE_CLOUD_ZONE"), "zone name for target cloud platform. defaults to GCP or set env GOOGLE_CLOUD_ZONE")
+	cmdImage.AddCommand(imageCreateCommand())
+	cmdImage.AddCommand(imageListCommand())
+	cmdImage.AddCommand(imageDeleteCommand())
+	cmdImage.AddCommand(imageResizeCommand())
+	cmdImage.AddCommand(imageSyncCommand())
+	return cmdImage
+}
+
+func imageCreateCommand() *cobra.Command {
+	var (
+		config, pkg, imageName string
+		args, mounts           []string
+		nightly                bool
+	)
+
+	var cmdImageCreate = &cobra.Command{
+		Use:   "create",
+		Short: "create nanos image from ELF",
+		Run:   imageCreateCommandHandler,
+	}
+
+	cmdImageCreate.PersistentFlags().StringVarP(&config, "config", "c", "", "ops config file")
+	cmdImageCreate.PersistentFlags().StringVarP(&pkg, "package", "p", "", "ops package name")
+	cmdImageCreate.PersistentFlags().StringArrayVarP(&args, "args", "a", nil, "command line arguments")
+	cmdImageCreate.PersistentFlags().StringArrayVar(&mounts, "mounts", nil, "mount <volume_id:mount_path>")
+	cmdImageCreate.PersistentFlags().BoolVarP(&nightly, "nightly", "n", false, "nightly build")
+
+	cmdImageCreate.PersistentFlags().StringVarP(&imageName, "imagename", "i", "", "image name")
+	return cmdImageCreate
+}
+
 func imageCreateCommandHandler(cmd *cobra.Command, args []string) {
 	provider, _ := cmd.Flags().GetString("target-cloud")
 	config, _ := cmd.Flags().GetString("config")
@@ -49,27 +92,12 @@ func imageCreateCommandHandler(cmd *cobra.Command, args []string) {
 		exitWithError("Please select on of the cloud platform in config. [onprem, aws, gcp, do, vsphere, vultr]")
 	}
 
-	if c.CloudConfig.Platform == "gcp" && len(c.CloudConfig.ProjectID) == 0 {
-		exitWithError("Please specify a cloud projectid in config")
-	}
-
 	if len(c.CloudConfig.BucketName) == 0 && c.CloudConfig.Platform != "onprem" {
 		exitWithError("Please specify a cloud bucket in config")
 	}
 
-	if len(pkg) > 0 {
-		c.Args = append(c.Args, cmdargs...)
-	} else {
-		if len(cmdargs) != 0 {
-			c.Program = cmdargs[0]
-		} else if len(c.Args) != 0 {
-			c.Program = c.Args[0]
-		} else {
-			exitWithError("Please mention program to run")
-		}
-	}
-
 	prepareImages(c)
+
 	// borrow BuildDir from config
 	bd := c.BuildDir
 	c.BuildDir = api.LocalVolumeDir
@@ -79,14 +107,15 @@ func imageCreateCommandHandler(cmd *cobra.Command, args []string) {
 	}
 	c.BuildDir = bd
 
-	p, err := getCloudProvider(provider, &c.CloudConfig)
+	p, ctx, err := getProviderAndContext(c, provider)
 	if err != nil {
 		exitWithError(err.Error())
 	}
-	ctx := api.NewContext(c, &p)
 
 	var keypath string
 	if len(pkg) > 0 {
+		c.Args = append(c.Args, cmdargs...)
+
 		expackage := downloadAndExtractPackage(pkg)
 
 		// load the package manifest
@@ -103,180 +132,40 @@ func imageCreateCommandHandler(cmd *cobra.Command, args []string) {
 		ctx = api.NewContext(c, &p)
 
 		keypath, err = p.BuildImageWithPackage(ctx, expackage)
-
+		if err != nil {
+			exitWithError(err.Error())
+		}
 	} else {
+		if len(cmdargs) != 0 {
+			c.Program = cmdargs[0]
+		} else if len(c.Args) != 0 {
+			c.Program = c.Args[0]
+		} else {
+			exitWithError("Please mention program to run")
+		}
+
 		setDefaultImageName(cmd, c)
 		keypath, err = p.BuildImage(ctx)
+		if err != nil {
+			exitWithError(err.Error())
+		}
 	}
 
+	err = p.CreateImage(ctx, keypath)
 	if err != nil {
 		exitWithError(err.Error())
 	}
 
-	if c.CloudConfig.Platform == "vultr" {
-		do := p.(*api.Vultr)
-		err = do.Storage.CopyToBucket(c, keypath)
-		if err != nil {
-			exitWithError(err.Error())
-		}
-
-		err = do.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("do image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
-	if c.CloudConfig.Platform == "do" {
-		do := p.(*api.DigitalOcean)
-		err = do.Storage.CopyToBucket(c, keypath)
-		if err != nil {
-			exitWithError(err.Error())
-		}
-
-		err = do.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("do image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
-	if c.CloudConfig.Platform == "gcp" {
-		gcloud := p.(*api.GCloud)
-		err = gcloud.Storage.CopyToBucket(c, keypath)
-		if err != nil {
-			exitWithError(err.Error())
-		}
-
-		err = gcloud.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("gcp image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
-	if c.CloudConfig.Platform == "aws" {
-		aws := p.(*api.AWS)
-
-		// verify we can even use the vm importer
-		api.VerifyRole(ctx, c.CloudConfig.BucketName)
-
-		err = aws.Storage.CopyToBucket(c, keypath)
-		if err != nil {
-			exitWithError(err.Error())
-		}
-
-		err = aws.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("aws image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
-	if c.CloudConfig.Platform == "vsphere" {
-		vsphere := p.(*api.Vsphere)
-		err = vsphere.Storage.CopyToBucket(c, keypath)
-		if err != nil {
-			exitWithError(err.Error())
-		}
-
-		err = vsphere.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("vsphere image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
-	if c.CloudConfig.Platform == "openstack" {
-		os := p.(*api.OpenStack)
-
-		err = os.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("openstack image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
-	if c.CloudConfig.Platform == "azure" {
-		azure := p.(*api.Azure)
-
-		err = azure.Storage.CopyToBucket(c, keypath)
-		if err != nil {
-			exitWithError(err.Error())
-		}
-
-		err = azure.CreateImage(ctx)
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("azure image '%s' created...\n", c.CloudConfig.ImageName)
-		}
-	}
-
+	fmt.Printf("%s image '%s' created...\n", provider, c.CloudConfig.ImageName)
 }
 
-func imageCreateCommand() *cobra.Command {
-	var (
-		config, pkg, imageName string
-		args, mounts           []string
-		nightly                bool
-	)
-
-	var cmdImageCreate = &cobra.Command{
-		Use:   "create",
-		Short: "create nanos image from ELF",
-		Run:   imageCreateCommandHandler,
+func imageListCommand() *cobra.Command {
+	var cmdImageList = &cobra.Command{
+		Use:   "list",
+		Short: "list images from provider",
+		Run:   imageListCommandHandler,
 	}
-
-	cmdImageCreate.PersistentFlags().StringVarP(&config, "config", "c", "", "ops config file")
-	cmdImageCreate.PersistentFlags().StringVarP(&pkg, "package", "p", "", "ops package name")
-	cmdImageCreate.PersistentFlags().StringArrayVarP(&args, "args", "a", nil, "command line arguments")
-	cmdImageCreate.PersistentFlags().StringArrayVar(&mounts, "mounts", nil, "mount <volume_id:mount_path>")
-	cmdImageCreate.PersistentFlags().BoolVarP(&nightly, "nightly", "n", false, "nightly build")
-
-	cmdImageCreate.PersistentFlags().StringVarP(&imageName, "imagename", "i", "", "image name")
-	return cmdImageCreate
-}
-
-// only targets local images today
-func imageResizeCommandHandler(cmd *cobra.Command, args []string) {
-
-	var c *api.Config
-	c = api.NewConfig()
-	AppendGlobalCmdFlagsToConfig(cmd.Flags(), c)
-
-	zone, _ := cmd.Flags().GetString("zone")
-	if zone != "" {
-		c.CloudConfig.Zone = zone
-	}
-
-	provider, _ := cmd.Flags().GetString("target-cloud")
-	p, err := getCloudProvider(provider, &c.CloudConfig)
-	if err != nil {
-		exitWithError(err.Error())
-	}
-	ctx := api.NewContext(c, &p)
-
-	err = p.ResizeImage(ctx, args[0], args[1])
-	if err != nil {
-		exitWithError(err.Error())
-	}
-}
-
-func imageResizeCommand() *cobra.Command {
-	var cmdImageResize = &cobra.Command{
-		Use:   "resize <image_name> <new_size>",
-		Short: "resize image",
-		Run:   imageResizeCommandHandler,
-		Args:  cobra.MinimumNArgs(2),
-	}
-	return cmdImageResize
+	return cmdImageList
 }
 
 func imageListCommandHandler(cmd *cobra.Command, args []string) {
@@ -304,20 +193,22 @@ func imageListCommandHandler(cmd *cobra.Command, args []string) {
 	}
 }
 
-func imageListCommand() *cobra.Command {
-	var cmdImageList = &cobra.Command{
-		Use:   "list",
-		Short: "list images from provider",
-		Run:   imageListCommandHandler,
+func imageDeleteCommand() *cobra.Command {
+	var cmdImageDelete = &cobra.Command{
+		Use:   "delete <image_name>",
+		Short: "delete images from provider",
+		Run:   imageDeleteCommandHandler,
+		Args:  cobra.MinimumNArgs(1),
 	}
-	return cmdImageList
+	return cmdImageDelete
 }
 
 func imageDeleteCommandHandler(cmd *cobra.Command, args []string) {
 	provider, _ := cmd.Flags().GetString("target-cloud")
+	config, _ := cmd.Flags().GetString("config")
+	config = strings.TrimSpace(config)
 
-	var c *api.Config
-	c = api.NewConfig()
+	c := unWarpConfig(config)
 	AppendGlobalCmdFlagsToConfig(cmd.Flags(), c)
 
 	zone, _ := cmd.Flags().GetString("zone")
@@ -334,27 +225,54 @@ func imageDeleteCommandHandler(cmd *cobra.Command, args []string) {
 	if err != nil {
 		exitWithError(err.Error())
 	}
+}
 
-	if c.CloudConfig.Platform == "azure" || provider == "azure" {
-		azure := p.(*api.Azure)
+func imageResizeCommand() *cobra.Command {
+	var cmdImageResize = &cobra.Command{
+		Use:   "resize <image_name> <new_size>",
+		Short: "resize image",
+		Run:   imageResizeCommandHandler,
+		Args:  cobra.MinimumNArgs(2),
+	}
+	return cmdImageResize
+}
 
-		err = azure.Storage.DeleteFromBucket(c, args[0]+".vhd")
-		if err != nil {
-			exitWithError(err.Error())
-		} else {
-			fmt.Printf("\nImage %s deleted successfully", args[0])
-		}
+// only targets local images today
+func imageResizeCommandHandler(cmd *cobra.Command, args []string) {
+	config, _ := cmd.Flags().GetString("config")
+	config = strings.TrimSpace(config)
+
+	c := unWarpConfig(config)
+	AppendGlobalCmdFlagsToConfig(cmd.Flags(), c)
+
+	zone, _ := cmd.Flags().GetString("zone")
+	if zone != "" {
+		c.CloudConfig.Zone = zone
+	}
+
+	provider, _ := cmd.Flags().GetString("target-cloud")
+	p, err := getCloudProvider(provider, &c.CloudConfig)
+	if err != nil {
+		exitWithError(err.Error())
+	}
+	ctx := api.NewContext(c, &p)
+
+	err = p.ResizeImage(ctx, args[0], args[1])
+	if err != nil {
+		exitWithError(err.Error())
 	}
 }
 
-func imageDeleteCommand() *cobra.Command {
-	var cmdImageDelete = &cobra.Command{
-		Use:   "delete <image_name>",
-		Short: "delete images from provider",
-		Run:   imageDeleteCommandHandler,
+func imageSyncCommand() *cobra.Command {
+	var sourceCloud string
+	var cmdImageSync = &cobra.Command{
+		Use:   "sync <image_name>",
+		Short: "sync image with from one provider to another",
+		Run:   imageSyncCommandHandler,
 		Args:  cobra.MinimumNArgs(1),
 	}
-	return cmdImageDelete
+	cmdImageSync.PersistentFlags().StringVarP(&sourceCloud, "source-cloud", "s", "onprem", "cloud platform [gcp, aws, do, vultr, onprem]")
+	return cmdImageSync
 }
 
 func imageSyncCommandHandler(cmd *cobra.Command, args []string) {
@@ -389,36 +307,4 @@ func imageSyncCommandHandler(cmd *cobra.Command, args []string) {
 	if err != nil {
 		exitWithError(err.Error())
 	}
-}
-
-func imageSyncCommand() *cobra.Command {
-	var sourceCloud string
-	var cmdImageSync = &cobra.Command{
-		Use:   "sync <image_name>",
-		Short: "sync image with from one provider to another",
-		Run:   imageSyncCommandHandler,
-		Args:  cobra.MinimumNArgs(1),
-	}
-	cmdImageSync.PersistentFlags().StringVarP(&sourceCloud, "source-cloud", "s", "onprem", "cloud platform [gcp, aws, do, vultr, onprem]")
-	return cmdImageSync
-}
-
-// ImageCommands provides image related command on GCP
-func ImageCommands() *cobra.Command {
-	var config, targetCloud, zone string
-	var cmdImage = &cobra.Command{
-		Use:       "image",
-		Short:     "manage nanos images",
-		ValidArgs: []string{"create", "list", "delete", "resize", "sync"},
-		Args:      cobra.OnlyValidArgs,
-	}
-	cmdImage.PersistentFlags().StringVarP(&config, "config", "c", "", "ops config file")
-	cmdImage.PersistentFlags().StringVarP(&targetCloud, "target-cloud", "t", "onprem", "cloud platform [gcp, aws, do, vultr, onprem]")
-	cmdImage.PersistentFlags().StringVarP(&zone, "zone", "z", os.Getenv("GOOGLE_CLOUD_ZONE"), "zone name for target cloud platform. defaults to GCP or set env GOOGLE_CLOUD_ZONE")
-	cmdImage.AddCommand(imageCreateCommand())
-	cmdImage.AddCommand(imageListCommand())
-	cmdImage.AddCommand(imageDeleteCommand())
-	cmdImage.AddCommand(imageResizeCommand())
-	cmdImage.AddCommand(imageSyncCommand())
-	return cmdImage
 }
