+++ lepton/azure_store.go
