+++ lepton/digital_ocean_store.go
