+++ lepton/vsphere_store.go
